# design/contracts/openapi/click-receiver.yaml
openapi: 3.1.0
info:
  title: click-receiver
  version: 1.0.0
  description: |
    The hot path. Route `/api/click-receiver`, **no JWT** (end users follow plain links).
    Both operations share one code path (spec §5.1):
      1. ad lookup (in-process cache 60 s → POSTGRES_READ_URL → POSTGRES_URL on a miss); unknown or inactive → 404
      2. dedup `SET click:dedup:<ad_id>:<user_id> <click_id> NX EX 600` (50 ms timeout; Redis error → fail open)
      3. accepted → produce ClickEvent (contracts/events/clicks.schema.json) to topic `clicks`, key = `ad_id`
         (or `ad_id#salt` when the ad is hot, spec §5.1.1), acks=all + idempotence, and **wait for the ack**
         before answering; accepted clicks feed the batched hot-ad counters in Redis
      4. produce failure/timeout (≤ 1.5 s) → DEL the dedup key (best effort) → 503, never "accepted"
    Every click response (302/200/404/503) carries `X-Click-Id`; 302/200 also carry `X-Click-Status` and
    `X-Click-Hot` (whether the receiver considered the ad hot; the key is salted when hot AND HOT_SALTING_ENABLED).
    Latency target: p95 < 100 ms.
servers:
  - url: /api/click-receiver
paths:
  /click/{ad_id}:
    get:
      operationId: clickRedirect
      summary: Record a click and redirect the browser to the ad's redirect_url
      parameters:
        - name: ad_id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
        - name: user_id
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      responses:
        "302":
          description: Accepted or duplicate; redirect to the advertiser
          headers:
            Location:
              required: true
              schema: { type: string }
              description: The ad's redirect_url
            X-Click-Status:
              required: true
              schema: { type: string, enum: [accepted, duplicate] }
            X-Click-Id:
              required: true
              schema: { type: string, format: uuid }
            X-Click-Hot:
              required: true
              schema: { type: string, enum: ["true", "false"] }
        "404":
          description: Unknown or inactive ad; nothing recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Missing/invalid user_id or ad_id
        "503":
          description: Click could not be durably recorded (Kafka ack failed/timed out); no redirect, the client may retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /clicks:
    post:
      operationId: recordClick
      summary: Record a click and return the redirect target as JSON (for SPAs and tests)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClickRequest" }
      responses:
        "200":
          description: Accepted or duplicate
          headers:
            X-Click-Status:
              required: true
              schema: { type: string, enum: [accepted, duplicate] }
            X-Click-Id:
              required: true
              schema: { type: string, format: uuid }
            X-Click-Hot:
              required: true
              schema: { type: string, enum: ["true", "false"] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClickResult" }
        "404":
          description: Unknown or inactive ad; nothing recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Invalid body
        "503":
          description: Click could not be durably recorded; the client may retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /hot-ads:
    get:
      operationId: listHotAds
      summary: This receiver's current view of hot ads (refreshed from Redis every ~2 s)
      responses:
        "200":
          description: Hot ads
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HotAds" }
components:
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
    ClickRequest:
      type: object
      required: [ad_id, user_id]
      properties:
        ad_id: { type: integer, minimum: 1 }
        user_id: { type: string, minLength: 1, maxLength: 128 }
    ClickResult:
      type: object
      required: [click_id, status, ad_id, redirect_url, clicked_at, hot]
      properties:
        click_id: { type: string, format: uuid }
        status: { type: string, enum: [accepted, duplicate] }
        ad_id: { type: integer }
        redirect_url: { type: string }
        clicked_at: { type: string, format: date-time }
        hot: { type: boolean, description: "Same as X-Click-Hot" }
    HotAd:
      type: object
      required: [ad_id, permanent, marks, hot_until]
      properties:
        ad_id: { type: integer }
        permanent: { type: boolean, description: "marks >= HOT_PERMANENT_AFTER_MARKS" }
        marks: { type: integer, minimum: 0, description: "Times the ad has been marked hot" }
        hot_until:
          type: [string, "null"]
          format: date-time
          description: When the temporary marking expires unless renewed; null for permanent ads without a current marking
    HotAds:
      type: object
      required: [refreshed_at, threshold_clicks_10m, salt_buckets, salting_enabled, items]
      properties:
        refreshed_at: { type: [string, "null"], format: date-time, description: "Last successful refresh from Redis" }
        threshold_clicks_10m: { type: integer }
        salt_buckets: { type: integer }
        salting_enabled: { type: boolean }
        items:
          type: array
          items: { $ref: "#/components/schemas/HotAd" }